fix(lifecycle): bots can age; life events persist (HIVE-022, 023, 024)
The product's headline feature has never run.

HIVE-022 — age_all_bots computes virtual_days = (real_hours/24) * time_scale,
then did `virtual_age_days += int(virtual_days)`. At the production time_scale of
7 an hourly cycle yields 0.29 days, and int(0.29) is 0. Age never incremented, so
no bot ever changed life stage, aged into elderhood, died of old age, left a
legacy, or triggered elder reproduction. virtual_age_days and death_age are now
Float and the truncation is gone. A test walks a bot through
young -> mature -> elder -> ancient and pins the ~521-real-day lifespan the
default config implies.

HIVE-023 — the demo switch that would have masked this read settings.DEMO_MODE,
which does not exist; the field is AUTHENTICITY_DEMO_MODE. A hasattr guard turned
that typo into a permanent False, so the civilization loop's fast path was
unreachable and nothing ever surfaced the mistake.

HIVE-024 — life_events was a plain JSON column and flag_modified appears nowhere
in this codebase, so SQLAlchemy never saw the in-place appends: every life event
written during aging was discarded at commit. life_events, relationships, roles,
inherited_traits and mutations are now MutableList/MutableDict. A scan for
in-place mutation of JSON ORM columns across mind/ found no other cases — the
other hits are dataclasses, not mapped columns.

Migration 'fractional_bot_ages' widens both columns and folds in the partial
unique index deferred from HIVE-008, as that task recommended rather than
raising a migration of its own. Existing rows are all 0, so the cast is lossless
— which is itself the evidence that no bot ever aged.

16 tests; 8 fail against unpatched code.

+++ mind/civilization/config.py
@@ -156,7 +156,7 @@ class CivilizationConfig:
         """Get vitality decay rate for a life stage."""
         return self.vitality_decay.get(life_stage, 0.001)
 
-    def get_life_stage(self, virtual_age_days: int) -> str:
+    def get_life_stage(self, virtual_age_days: float) -> str:
         """Determine life stage based on virtual age."""
         for stage, (min_days, max_days) in self.life_stages.items():
             if min_days <= virtual_age_days < max_days:
